==> 05-Bacon.Rmd <==
---
title: "Adipokine EWAS - Script 5: Correcting for bias and inflation in test statistics with bacon"
author: "Lucy Sinke"
output: html_document
---

```{r echo=FALSE}
rm(list=ls())
```

```{r, setup, include=FALSE}
knitr::opts_knit$set(root.dir = '/exports/molepi/users/ljsinke/LLS/Adi_Scripts')
```

***

# Setup

Load packages

```{r message=FALSE}
library(tidyverse)
library(bacon)
```


***

# Adiponectin EWAS - Base Model

***

### Load Data

This contains data on:

* `cpg` - probe ID
* `beta_` - effect size reported by each cohort
* `SE_` - standard error reported by each cohort
* `p_`- unadjusted p-value reported by each cohort
* `N_` - sample size of each cohort

```{r}
load('../Adi_Data/Processing/ALL-2-PRUNE_ln-adi_base.Rdata')
limma_base[1:5, 1:5]
```

***

### Bacon

Create functions

```{r}
run_bacon <- function(cohort){
  
  print("---------------------------------------------------------")
  
  # Subset data frame
  df <- limma_base %>% 
    dplyr::select(
      cpg,
      beta = paste0("beta_", cohort),
      se = paste0("SE_", cohort),
      N = paste0("N_", cohort)
    ) %>% 
    filter(!is.na(beta) & !is.na(se))
  
  print(paste0("For cohort ", cohort, 
               " we have data on ", nrow(df), " CpGs..."))
  
  # Save values
  cpg <- df$cpg
  num <- df$N
  es <- as.numeric(df$beta)
  se <- as.numeric(df$se)
  
  # Run bacon
  bc <- bacon(NULL, es, se)
  print(bc)
  
  # Save output
  tstat <- tstat(bc)
  p <- pval(bc) 
  coef <- es(bc)
  se <- se(bc)
  
  # Rerun bacon
  bc <- bacon(NULL, coef, se)
  print(bc)
  
  # Save output
  assign(paste0("bacon_", cohort), data.frame(
    cpg, coef, se, p, tstat, num
  ), envir=.GlobalEnv)
  
  name_list <- c(
    "cpg", 
    paste0("beta_", cohort), 
    paste0("SE_", cohort),
    paste0("p_", cohort),
    paste0("tstat_", cohort),
    paste0("N_", cohort))
  
  assign(paste0("bacon_",cohort),
         setNames(get(paste0("bacon_",cohort)), name_list),
         envir=.GlobalEnv)
}
```

***

LLS

```{r}
run_bacon("lls")
```

***

KORA

```{r}
run_bacon("kora")
```

***

TwinsUK

```{r}
run_bacon("tuk")
```

***

LifeLines Deep

```{r}
run_bacon("lld")
```

***

SHIP

```{r}
run_bacon("ship")
```

***

GOLDN

```{r}
run_bacon("goldn")
```

***

Merge

```{r}
bacon_df <- full_join(
  bacon_lls, bacon_kora, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_tuk, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_lld, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_ship, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_goldn, by = "cpg"
)

#bacon_df <- bacon_df %>% filter(!abs(beta_lld) > 1)

dim(bacon_df)
head(bacon_df)
```

***

Save

```{r}
save(bacon_df, file='../Adi_Data/Processing/ALL-3-BACON_ln-adi_base.Rdata')
```


***

# Adiponectin EWAS - Smoking Sensitivity Analysis

***

### Load Data

This contains data on:

* `cpg` - probe ID
* `beta_` - effect size reported by each cohort
* `SE_` - standard error reported by each cohort
* `p_`- unadjusted p-value reported by each cohort
* `N_` - sample size of each cohort

```{r}
load('../Adi_Data/Processing/ALL-2-PRUNE_ln-adi_ext1.Rdata')
limma_base[1:5, 1:5]
```

***

### Bacon

***

LLS

```{r}
run_bacon("lls")
```

***

KORA

```{r}
run_bacon("kora")
```

***

TwinsUK

```{r}
run_bacon("tuk")
```

***

LifeLines Deep

```{r}
run_bacon("lld")
```

***

SHIP

```{r}
run_bacon("ship")
```

***

GOLDN

```{r}
run_bacon("goldn")
```


***

Merge

```{r}
bacon_df <- full_join(
  bacon_lls, bacon_kora, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_tuk, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_lld, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_ship, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_goldn, by = "cpg"
)

#bacon_df <- bacon_df %>% filter(!abs(beta_lld) > 1)

dim(bacon_df)
head(bacon_df)
```

***

Save

```{r}
save(bacon_df, file='../Adi_Data/Processing/ALL-3-BACON_ln-adi_ext1.Rdata')
```

***

# Adiponectin EWAS - BMI Sensitivity Analysis

***

### Load Data

This contains data on:

* `cpg` - probe ID
* `beta_` - effect size reported by each cohort
* `SE_` - standard error reported by each cohort
* `p_`- unadjusted p-value reported by each cohort
* `N_` - sample size of each cohort

```{r}
load('../Adi_Data/Processing/ALL-2-PRUNE_ln-adi_ext2.Rdata')
limma_base[1:5, 1:5]
```

***

### Bacon

***

LLS

```{r}
run_bacon("lls")
```

***

KORA

```{r}
run_bacon("kora")
```

***

TwinsUK

```{r}
run_bacon("tuk")
```

***

LifeLines Deep

```{r}
run_bacon("lld")
```

***

SHIP

```{r}
run_bacon("ship")
```

***

GOLDN

```{r}
run_bacon("goldn")
```

***

Merge

```{r}
bacon_df <- full_join(
  bacon_lls, bacon_kora, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_tuk, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_lld, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_ship, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_goldn, by = "cpg"
)

#bacon_df <- bacon_df %>% filter(!abs(beta_lld) > 1)

dim(bacon_df)
head(bacon_df)
```

***

Save

```{r}
save(bacon_df, file='../Adi_Data/Processing/ALL-3-BACON_ln-adi_ext2.Rdata')
```

***



# Adiponectin EWAS - Epidish sensitivity analysis

***

### Load Data

This contains data on:

* `cpg` - probe ID
* `beta_` - effect size reported by each cohort
* `SE_` - standard error reported by each cohort
* `p_`- unadjusted p-value reported by each cohort
* `N_` - sample size of each cohort

```{r}
load('../Adi_Data/Processing/ALL-2-PRUNE_ln-adi_ext3.Rdata')
limma_base[1:5, 1:5]
```

***

### Bacon

***

LLS

```{r}
run_bacon("lls")
```

***

KORA

```{r}
run_bacon("kora")
```

***

LifeLines Deep

```{r}
run_bacon("lld")
```

***

SHIP

```{r}
run_bacon("ship")
```

***

Merge

```{r}
bacon_df <- full_join(
  bacon_lls, bacon_kora, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_tuk, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_lld, by = "cpg"
)

bacon_df <- full_join(
  bacon_df, bacon_ship, by = "cpg"
)

#bacon_df <- bacon_df %>% filter(!abs(beta_lld) > 1)

dim(bacon_df)
head(bacon_df)
```

***

Save

```{r}
save(bacon_df, file='../Adi_Data/Processing/ALL-3-BACON_ln-adi_ext3.Rdata')
```

***
