==> 10_DMR.Rmd <==
---
title: "Adipokine EWAS - Script 10: Identifying distinct loci"
author: "Lucy Sinke"
output: html_document
---

```{r echo=FALSE}
rm(list=ls())
```

***

# Setup

Load packages

```{r warning=FALSE, message=FALSE}
library(tidyverse)
library(ggrepel)
library(GenomicRanges)
library(ggpubr)
library(DNAmArray)
```

***

Load data

```{r}
load("../Adi_Data/Results/AdiTop_2024.Rdata")
adi_top <- top
head(adi_top)
dim(adi_top)
```

***

Make DMR data frame

```{r}
adi_top <- adi_top %>% 
  dplyr::select(
    chromosome = cpg_chr_hg19,
    start = cpg_start_hg19,
    padj = base_meta_padj_fdr,
    het = base_het_p
  ) %>% filter(!is.na(chromosome) & chromosome!='chrX')
```


```{r}
adi_top <- adi_top %>% mutate(
    padj = ifelse(is.na(padj), 1, padj))

#res <- res %>% mutate(
#  chromosome = as.numeric(sub('chr', '', chromosome)))

adi_top <- adi_top %>% mutate(
    crit = ifelse(padj <= 0.05, 1, 0))
```


```{r}
adi_top <- adi_top %>% filter(!is.na(crit) &
                        !is.na(chromosome))
```

```{r}
adi_top <- adi_top %>% arrange(chromosome, start) %>% 
  dplyr::select(chromosome, start, crit)
```

```{r}
chromosome=1:22
MAXIMUM_REGION_LENGTH = 1000
mismatches = 3
chr_list <- 1:22
```

Chromosome 1


```{r}
for(x in chr_list){

chr1 = adi_top[adi_top[,1]==x,]
chr1 <- chr1 %>% arrange(start)
chr.final = data.frame(
  coord = chr1$start,
  crit = chr1$crit
)

last_coordinate = length( chr.final$crit )
next_coordinate = 0

for (i in 1:(last_coordinate-1)) {
      if ( i>=next_coordinate ) {
        if (chr.final$crit[ i ]==1) {
          start_location = chr.final$coord[ i ]
          last_visited_crit_loc = start_location
          sum_of_ones = 1
          number_of_items = 1
          
          # start crawling loop
          for (j in (i+1):last_coordinate ) {
            if (chr.final$coord[ j ] > (last_visited_crit_loc + MAXIMUM_REGION_LENGTH)) { break }
            if((number_of_items-sum_of_ones)>mismatches) { break }   #Number of mismatches
            number_of_items = number_of_items + 1
            if (chr.final$crit[j]==1) { 
              last_visited_crit_loc = chr.final$coord[ j ]
              sum_of_ones = sum_of_ones + 1 
            }
          }
          
          # now check if the result is good enough
          if (sum_of_ones>=3) {
            last_one=i+number_of_items-1
            for (k in (i+number_of_items-1):1) {
              if ( chr.final$crit[k] == 0 ) {
                last_one = last_one - 1
                number_of_items = number_of_items - 1
              }
              else {
                break
              }
            }
            cat(x, ';',start_location,";",chr.final$coord[last_one],";",sum_of_ones/number_of_items,"\n")
            next_coordinate = last_one + 1
          }
        }
      }
}
}
```

***



***
