==> 12_EWAS-Enrichment.Rmd <==
---
title: "Adipokine EWAS - Script 12: EWAS Enrichment"
author: "Lucy Sinke"
date: "10/31/2022"
output: html_document
---

```{r echo=FALSE}
rm(list=ls())
```

```{r, setup, include=FALSE}
knitr::opts_knit$set(root.dir = '/exports/molepi/users/ljsinke/LLS/Adi_Scripts/')
```

# Setup

Load packages

```{r}
library(tidyverse)
library(MASS)
```

Make a list of eligible CpGs:

* Measured with results in EWAS
* Not removed due to high heterogeneity
* Not removed due to IDOL extended sensitivity analyses

```{r}
load("../Adi_Data/Results/Adi_EWAS_Results-Full.Rdata")
load("../Adi_Data/Results/AdiTop.Rdata")
```

Save significant CpGs

```{r}
sig_cpgs <- top$cpg
length(sig_cpgs)
```

***

# EWAS Catalog

Load EWAS catalog (2,059,897 associations)

```{r}
ewas_res <- read_tsv('../IL6_Data/EWAS/ewascatalog-results.txt.gz')
ewas_stu <- read_tsv('../IL6_Data/EWAS/ewascatalog-studies.txt.gz')
```

Merge results with studies as one data frame

```{r}
ewas_cat <- left_join(ewas_res, ewas_stu, by='StudyID')
rm(list=c('ewas_res', 'ewas_stu'))
```

Remove those with N under 500

```{r}
ewas_cat <- ewas_cat %>% filter(N>=500)
```

Remove those without PMID (1,743,454 associations remain)

```{r}
ewas_cat <- ewas_cat[!is.na(as.numeric(ewas_cat$PMID)),]
```

Look at tissues & keep blood EWAS only (whole blood, blood, or leukocytes - 1,107,887 associations remain):

```{r}
#table(ewas_cat$Tissue)

ewas_cat <- ewas_cat %>% filter(Tissue %in% c("blood", 
                                                "Blood", 
                                                "CD4+ T-cells, whole blood",
                                                "CD4+ T-cells, Whole blood",
                                                "Cord blood, whole blood",
                                                "Leukocytes",
                                                "Peripheral blood",
                                                "whole blood",
                                                "Whole blood",
                                                "Whole Blood",
                                                "Whole blood, breast tissue",
                                                "Whole blood, CD4+ T-cells, CD14+ monocytes",
                                                "Whole blood, CD4+ T cells",
                                                "Whole blood, CD4+ T cells, CD14+ monocytes",
                                                "Whole blood, cord blood",
                                                "Whole blood, heel prick blood spots"
))
```

Keep only EWAS in adults

```{r}
table(ewas_cat$Age)

ewas_cat <- ewas_cat %>% filter(Age %in% c(
  "Adults", 
  "Adults (18-65 years)", 
  "Geriatrics"))
```

Make a key of traits collapsing differently worded similar traits (e.g. BMI, bmi, body mass index, Body Mass Index etc.)

```{r}
#write_csv(as.data.frame(table(ewas_cat$Trait)), file='../IL6_Data/EWAS/EWAS_cat_Trait.csv')
```

Load key for traits

```{r}
trait_key <- read_csv('../IL6_Data/EWAS/EWAS_cat_Trait.csv')
ewas_cat <- left_join(ewas_cat, trait_key, by = 'Trait')
```

Keep only our CpGs

```{r}
ewas_cat <- ewas_cat %>% filter(CpG %in% adi_results$cpg)
```

Combine

```{r}
ewas_full <- ewas_cat %>% dplyr::select(cpg = CpG, Trait = Trait_Simple)
dim(ewas_full)
length(unique(ewas_full$cpg))
```

***

# EWAS Atlas

Load EWAS Atlas (617,084)

```{r}
ewas_atlas <- read_tsv('../IL6_Data/EWAS/EWAS_Atlas_associations.tsv')
ewas_stu <- read_tsv('../IL6_Data/EWAS/EWAS_Atlas_cohorts.tsv')
ewas_stu <- ewas_stu %>% dplyr::select(study_ID, tissue, mean_age, sample_size) 
ewas_stu <- ewas_stu[!duplicated(ewas_stu$study_ID),]
ewas_atlas <- left_join(ewas_atlas, ewas_stu, by='study_ID')
rm('ewas_stu')
```

Remove those with N under 500

```{r}
ewas_atlas <- ewas_atlas %>% filter(sample_size >= 500)
```

Remove those without PMID (562,488 associations remain)

```{r}
ewas_atlas <- ewas_atlas %>% filter(!is.na(as.numeric(PMID))) %>% filter(!PMID %in% ewas_cat$PMID)
```

Look at tissues & keep blood EWAS only (whole blood, blood, or leukocytes - 153,777 associations remain):

```{r}
table(ewas_atlas$tissue)

ewas_atlas <- ewas_atlas %>% filter(tissue %in% c(
  "blood",
  "blood spot",
  "bloodspot",
  "buffy coat",
  "bufy coat",
  "leukocyte",
  "leukocytes",
  "peripheral blood",
  "white blood cells",
  "whole blood"
))
```

Keep only EWAS in adults (77,974)

```{r}
ewas_atlas <- ewas_atlas %>% filter(mean_age >= 18)
```

Collapse traits

```{r}
#write_csv(as.data.frame(table(ewas_atlas$trait)), file='../IL6_Data/EWAS/EWAS_atlas_Trait.csv')
```

Load key for traits

```{r}
trait_key <- read_csv('../IL6_Data/EWAS/EWAS_atlas_Trait.csv')
ewas_atlas <- left_join(ewas_atlas, trait_key, by = 'trait')
```

Keep only our CpGs

```{r}
ewas_atlas <- ewas_atlas %>% filter(probe_ID %in% adi_results$cpg)
dim(ewas_atlas)
length(unique(ewas_atlas$probe_ID))
```

Merge

```{r}
ewas_full <- rbind(ewas_full, ewas_atlas %>% dplyr::select(cpg = probe_ID, Trait=Trait_Simple))
dim(ewas_full)
length(unique(ewas_full$cpg))
```

***

# Enrichment

For each trait we want:

* To save the CpGs found in EWAS of that trait in a data frame 'cpg_trait'
* To create a new data frame from `cpg_list` with `trait` as a binary indicator of if that CpG was found in that trait EWAS
* Perform logistic regression for both Bonfi and FDR CpGs 
* Print the output

```{r}
i <- "BMI"

cpg_trait <- (ewas_full %>% filter(Trait == i))$cpg
n <- length(cpg_trait)

if(n >= 100) {df <- adi_results %>% 
  mutate(
    trait = ifelse(cpg %in% cpg_trait, 1, 0),
    sig = ifelse(cpg %in% sig_cpgs, 1, 0)
  )
  nFound = sum(df$trait)
  nSig = sum((df %>% filter(sig == 1))$trait)

  trait_df <- df %>% filter(trait == 1)
  save(trait_df, file=paste0('../Adi_Data/EWAS/Trait/', i,     
  '.Rdata'))

  x <- glm(trait ~ sig, family=binomial, data=df)
  out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
  names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
  out <- as.data.frame(t(out))
  out$Trait = i
  out$nCpG = n
  out$nFound = nFound
  out$nSig = nSig
  out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p, nCpG, nFound, nSig)
  out$padj <- p.adjust(out$p, method='fdr')
  res <- out
}
```

```{r}
trait_list <- unique(ewas_full$Trait)

for(i in trait_list){
 cpg_trait <- (ewas_full %>% filter(Trait == i))$cpg
 n <- length(cpg_trait)

if (n >= 100){
  df <- adi_results %>% 
    mutate(
    trait = ifelse(cpg %in% cpg_trait, 1, 0),
    sig = ifelse(cpg %in% sig_cpgs, 1, 0)
  )
  nFound = sum(df$trait)
  nSig = sum((df %>% filter(sig == 1))$trait)
  
  trait_df <- df %>% filter(trait == 1)
  save(trait_df, file=paste0('../Lep_Data/EWAS/Trait/', i,   
  '.Rdata'))

  x <- glm(trait ~ sig, family=binomial, data=df)
  out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
  names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
  out <- as.data.frame(t(out))
  out$Trait = i
  out$nCpG = n
  out$nFound = nFound
  out$nSig = nSig
  out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p, nCpG, nFound, nSig) 
  out$padj <- p.adjust(out$p, method='fdr')
  res <- rbind(out, res)
}    
}

write_csv(res %>% arrange(p, OR),file="../Adi_Data/EWAS/EWAS_preBMI.csv")
```

***

# Filter BMI

```{r}
top$ext2_meta_padj_fdr <- p.adjust(top$ext2_meta_p, method='fdr')
nrow(top %>% filter(ext2_meta_padj_fdr <= 0.05))
```


```{r}
ggplot(top, aes(x=base_meta_beta, 
                 y=ext2_meta_beta)) +
  geom_hline(aes(yintercept=0),
             color="#1B2021") +
  geom_vline(aes(xintercept=0),
             color="#1B2021") +
  geom_abline(intercept = 0, slope = 1, color = 'grey40', linetype='dashed') +
  geom_point(color = ifelse(top$ext2_meta_padj_fdr > 0.05, "#B3AF8F", "white"),
             alpha = ifelse(top$ext2_meta_padj_fdr > 0.05, 1, 0)) +
  geom_point(color = ifelse(top$ext2_meta_padj_fdr <= 0.05, "#237194", "white"),
             alpha = ifelse(top$ext2_meta_padj_fdr <= 0.05, 1, 0)) +
  stat_cor(data = top %>% filter(ext2_meta_padj_fdr <= 0.05),
           color = "#237194",
           p.accuracy = 0.001, 
           r.accuracy=0.01) +
  xlab('Effect Size (Base)') +
  ylab('Effect Size (BMI)') +
  ggtitle('') +
  theme(
    axis.text = element_text(
       size=9, 
       color="#1B2021"),
    axis.title = element_text(
      size=11, 
      hjust=0.5, 
      color="#1B2021"),
    plot.title = element_text(
      size=16, 
      hjust=0.5,
      face="bold", 
      color = "#548687"),
    panel.background = element_rect(
      fill="white"),
    panel.border = element_rect(
      color="#1B2021",
      fill=NA),
    panel.grid.major = element_line(
      color="grey95"),
    panel.grid.minor = element_line(
      color="grey95"),
    plot.background = element_rect(
      fill="white"),
    legend.position="none")
```


```{r}
bmi_cpgs <- (top %>% filter(ext2_meta_padj_fdr<=0.05))$cpg
```

***

# Enrichment

For each trait we want:

* To save the CpGs found in EWAS of that trait in a data frame 'cpg_trait'
* To create a new data frame from `cpg_list` with `trait` as a binary indicator of if that CpG was found in that trait EWAS
* Perform logistic regression for both Bonfi and FDR CpGs 
* Print the output

```{r}
i <- "BMI"

cpg_trait <- (ewas_full %>% filter(Trait == i))$cpg
n <- length(cpg_trait)

if(n >= 100) {df <- adi_results %>% 
  mutate(
    trait = ifelse(cpg %in% cpg_trait, 1, 0),
    sig = ifelse(cpg %in% bmi_cpgs, 1, 0)
  )
  nFound = sum(df$trait)
  nSig = sum((df %>% filter(sig == 1))$trait)

  trait_df <- df %>% filter(trait == 1)
  save(trait_df, file=paste0('../Adi_Data/EWAS/Trait/', i,     
  '_postBMI.Rdata'))

  x <- glm(trait ~ sig, family=binomial, data=df)
  out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
  names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
  out <- as.data.frame(t(out))
  out$Trait = i
  out$nCpG = n
  out$nFound = nFound
  out$nSig = nSig
  out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p, nCpG, nFound, nSig)
  out$padj <- p.adjust(out$p, method='fdr')
  res <- out
}
```

```{r}
trait_list <- unique(ewas_full$Trait)

for(i in trait_list){
 cpg_trait <- (ewas_full %>% filter(Trait == i))$cpg
 n <- length(cpg_trait)

if (n >= 100){
  df <- adi_results %>% 
    mutate(
    trait = ifelse(cpg %in% cpg_trait, 1, 0),
    sig = ifelse(cpg %in% bmi_cpgs, 1, 0)
  )
  nFound = sum(df$trait)
  nSig = sum((df %>% filter(sig == 1))$trait)
  
  trait_df <- df %>% filter(trait == 1)
  save(trait_df, file=paste0('../Adi_Data/EWAS/Trait/', i,   
  '_postBMI.Rdata'))

  x <- glm(trait ~ sig, family=binomial, data=df)
  out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
  names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
  out <- as.data.frame(t(out))
  out$Trait = i
  out$nCpG = n
  out$nFound = nFound
  out$nSig = nSig
  out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p, nCpG, nFound, nSig) 
  out$padj <- p.adjust(out$p, method='fdr')
  res <- rbind(out, res)
}    
}

write_csv(res %>% arrange(p, OR),file="../Adi_Data/EWAS/EWAS_postBMI.csv")
```

***

Save

```{r}
top <- top %>% filter(ext2_meta_padj_fdr <= 0.05)
save(top, file='../Adi_Data/Results/AdiTop_2024_postBMI.Rdata')
```
