==> 13_Chrom_Enrichment.Rmd <==
---
title: "Adipokine EWAS - Script 13: Chromatin state enrichment"
author: "Lucy Sinke"
output: html_document
---

```{r echo=FALSE}
rm(list=ls())
```

***

# Setup

Load packages

```{r warning=FALSE, message=FALSE}
library(tidyverse)
library(ggrepel)
library(GenomicRanges)
library(ggpubr)
library(DNAmArray)
library(MASS)
```



***

```{r}
load("../Adi_Data/Results/Adi_EWAS_Results-Full.Rdata")
dim(adi_results)

load("../Adi_Data/Adiponectin_eqtm.Rdata")
top <- sig
```


```{r}
top <- read_tsv('cis_cpg.tsv')
```

***

# Enrichment tests

```{r}
res_road <- adi_results  %>% 
  mutate(sig = ifelse(adi_results$cpg %in% top$cpg, 1, 0),
         `1_TssA` = ifelse(E062 == "1_TssA", 1, 0),
         `2_TssAFlnk` = ifelse(E062 == "2_TssAFlnk", 1, 0),
         `3_TxFlnk` = ifelse(E062 == "3_TxFlnk", 1, 0),
         `4_Tx` = ifelse(E062 == "4_Tx", 1, 0),
         `5_TxWk` = ifelse(E062 == "5_TxWk", 1, 0),
         `6_EnhG` = ifelse(E062 == "6_EnhG", 1, 0),
         `7_Enh` = ifelse(E062 == "7_Enh", 1, 0),
         `8_ZNF/Rpts` = ifelse(E062 == "8_ZNF/Rpts", 1, 0),
         `9_Het` = ifelse(E062 == "9_Het", 1, 0),
         `10_TssBiv` = ifelse(E062 == "10_TssBiv", 1, 0),
         `11_BivFlnk` = ifelse(E062 == "11_BivFlnk", 1, 0),
         `12_EnhBiv` = ifelse(E062 == "12_EnhBiv", 1, 0),
         `13_ReprPC` = ifelse(E062 == "13_ReprPC", 1, 0),
         `14_ReprPCWk` = ifelse(E062 == "14_ReprPCWk", 1, 0),
         `15_Quies` = ifelse(E062 == "15_Quies", 1, 0)) 

head(res_road %>% dplyr::select(E062, `1_TssA`, `2_TssAFlnk`, `3_TxFlnk`,
                           `4_Tx`, `5_TxWk`, `6_EnhG`, `7_Enh`, 
                           `8_ZNF/Rpts`, `9_Het`, `10_TssBiv`, 
                           `11_BivFlnk`, `12_EnhBiv`, `13_ReprPC`, 
                           `14_ReprPCWk`, `15_Quies`))

res_road <- res_road %>% mutate(
  `1_TssA` = ifelse(is.na(`1_TssA`), 0, `1_TssA`),
  `2_TssAFlnk` = ifelse(is.na(`2_TssAFlnk`), 0, `2_TssAFlnk`),
  `3_TxFlnk` = ifelse(is.na(`3_TxFlnk`), 0, `3_TxFlnk`),
  `4_Tx` = ifelse(is.na(`4_Tx`), 0, `4_Tx`),
  `5_TxWk` = ifelse(is.na(`5_TxWk`), 0, `5_TxWk`),
  `6_EnhG` = ifelse(is.na(`6_EnhG`), 0, `6_EnhG`),
  `7_Enh` = ifelse(is.na(`7_Enh`), 0, `7_Enh`),
  `8_ZNF/Rpts` = ifelse(is.na(`8_ZNF/Rpts`), 0, `8_ZNF/Rpts`),
  `9_Het` = ifelse(is.na(`9_Het`), 0, `9_Het`),
  `10_TssBiv` = ifelse(is.na(`10_TssBiv`), 0, `10_TssBiv`),
  `11_BivFlnk` = ifelse(is.na(`11_BivFlnk`), 0, `11_BivFlnk`),
  `12_EnhBiv` = ifelse(is.na(`12_EnhBiv`), 0, `12_EnhBiv`),
  `13_ReprPC` = ifelse(is.na(`13_ReprPC`), 0, `13_ReprPC`),
  `14_ReprPCWk` = ifelse(is.na(`14_ReprPCWk`), 0, `14_ReprPCWk`),
  `15_Quies` = ifelse(is.na(`15_Quies`), 0, `15_Quies`),
)
```

GLM for each chromatin state

```{r}
x <- glm(`1_TssA` ~ sig, family=binomial, data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "1_TssA"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- out

x <- glm(`2_TssAFlnk` ~ sig, family=binomial, data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "2_TssAFlnk"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`3_TxFlnk` ~ sig, family=binomial, data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "3_TxFlnk"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`4_Tx` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "4_Tx"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`5_TxWk` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "5_TxWk"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`6_EnhG` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "6_EnhG"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`7_Enh` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "7_Enh"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`8_ZNF/Rpts` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "8_ZNF/Rpts"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`9_Het` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "9_Het"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`10_TssBiv` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "10_TssBiv"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`11_BivFlnk` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "11_BivFlnk"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`12_EnhBiv` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "12_EnhBiv"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`13_ReprPC` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "13_ReprPC"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)

x <- glm(`14_ReprPCWk` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "14_ReprPCWk"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res) 

x <- glm(`15_Quies` ~ sig, family=binomial  , data=res_road)
out <- c(coef(summary(x))[2,],exp(cbind(coef(x), confint.default(x)))[2,])
names(out) <- c('logOR', 'SE', 'z', 'p', 'OR', 'low_CI', 'upp_CI')
out <- as.data.frame(t(out))
out$Trait = "15_Quies"
out <- out %>% dplyr::select(Trait, OR, logOR, low_CI, upp_CI, z, p) 
res <- rbind(out, res)
```

```{r}
res$padj <- p.adjust(res$p, method='fdr')
```


Save results

```{r}
write_csv(res %>% arrange(p, OR),file="../Adi_Data/Chromatin_enrichment_eqtm.csv")
```

***
